refactor: move CertificateChain out of nexus into test-utils
